requirements.txt
Flask==2.0.2
openai==0.10.2
pytesseract==0.3.8
Pillow==8.4.0
tensorflow==2.16.2
opencv-python==4.5.3.56
numpy==1.19.5
joblib==1.1.0
